JMS Pool configuration options cleanup
Rename many configuration options to use simpler more understandable
names and remove legacy options that don't work as they should or are
not implemented in a fashion that would make them useful to most users
that would try and use them.  Update documentation for more clarity on
the way they affect the pool.